requirements.txt
pyusb==1.2.1
requests==2.31.0